Fix for constants;

+++ src/TokenByte.java
@@ -1,9 +1,9 @@
 @SuppressWarnings("WeakerAccess")
-public class TokenByte extends TokenConstant<Byte> {
+public class TokenByte extends TokenConstant<Short> {
 
     public TokenByte(String lexeme) {
         super(lexeme, ConstantType.BYTE);
-        setConstant(Byte.decode(lexeme.toLowerCase().replace('h', 'x')));
+        setConstant(Short.decode(lexeme.toLowerCase().replace('h', 'x')));
     }
 
 }
